change sql create format, revert back to single database instead of multi database with merging, change the cutoff point to Wednesday and Saturday.

+++ human_tracker/main.py
@@ -20,8 +20,9 @@ flags.DEFINE_boolean('tiny', False, 'yolo or yolo-tiny')
 flags.DEFINE_string('model', 'yolov4', 'yolov3 or yolov4')
 flags.DEFINE_string('video', './data/video/', 'path to input video or set to 0 for webcam')
 flags.DEFINE_string('rtsp_path', 'data/rtsp/rtsp_cam.xlsx', 'default rtsp camera path')
-flags.DEFINE_string('cam_db_path', '../reid/database', 'default cam database path')
-flags.DEFINE_string('merge_db_path', "../reid/database/merge", 'default merged reid database path, where all of the samples from cam are saved into the db with timestamp')
+flags.DEFINE_string('reid_db_path', "../reid/archive", 'default reid database path, where all of the samples from cam are saved into the db with timestamp')
+#flags.DEFINE_string('cam_db_path', '../reid/database', 'default cam database path')
+#flags.DEFINE_string('merge_db_path', "../reid/database/merge", 'default merged reid database path, where all of the samples from cam are saved into the db with timestamp')
 #flags.DEFINE_string('output', './outputs/', 'path to output video')
 flags.DEFINE_boolean('output', False, 'path to output video')
 flags.DEFINE_string('output_format', 'MJPG', 'codec used in VideoWriter when saving video to file')
@@ -30,7 +31,7 @@ flags.DEFINE_float('score', 0.50, 'score threshold')
 flags.DEFINE_boolean('dont_show', False, 'dont show video output')
 flags.DEFINE_boolean('info', False, 'show detailed info of tracked objects')
 flags.DEFINE_boolean('count', False, 'count objects being tracked on screen')
-flags.DEFINE_boolean('db', True, 'save information in database')
+flags.DEFINE_boolean('db', False, 'save information in database')
 flags.DEFINE_boolean('trajectory', False, 'draw historical trajectories on every tracked human')
 flags.DEFINE_integer('input_skip_frame', 8, 'number of frame to be skipped')
 flags.DEFINE_integer('db_skip_frame', 8, 'number of frame to be skipped')
@@ -40,52 +41,52 @@ flags.DEFINE_integer('parallel_ps', 2, 'number of human tracker process to run')
 flags.DEFINE_boolean('online', False, 'run online image extraction using rtsp')
 flags.DEFINE_boolean('reid', True, 'set to True to run with REID, set to False if new labelled data are needed to be recorded')
 
-def db_process(*args):
-    def signal_handler(sig, frame):
-        name = mp.current_process().name
-        print(str(name) + ': You pressed Ctrl+C!')
+# def db_process(*args):
+#     def signal_handler(sig, frame):
+#         name = mp.current_process().name
+#         print(str(name) + ': You pressed Ctrl+C!')
 
-        db_list = [] 
-        # args[0] is the camera list
-        # args[1] is the length of camera list
-        # args[2] is the cam path
-        # args[3] is the db merge path
-        # args[4] is the shared queue recording the database paths 
+#         db_list = [] 
+#         # args[0] is the camera list
+#         # args[1] is the length of camera list
+#         # args[2] is the cam path
+#         # args[3] is the db merge path
+#         # args[4] is the shared queue recording the database paths 
         
-        for i in range(args[1]):
-            if args[0]:
-                db_list.append(args[2] + "/Cam_" + str(args[0][i]) + ".db")
-        # finish gathering the db_paths, run merge.
-        print('Saving merge database..')
-        now = dt.datetime.now()
-        db_name = now.strftime("Reid_Interrputed_%Y%m%d.db")
-        db_filepath = os.path.join(args[3], db_name)
-        reid_db = ImageDB(db_name=db_filepath)
-        reid_db.delete_dbfile()
-        reid_db.create_table()
-        reid_db.merge_data(db_list)
+#         for i in range(args[1]):
+#             if args[0]:
+#                 db_list.append(args[2] + "/Cam_" + str(args[0][i]) + ".db")
+#         # finish gathering the db_paths, run merge.
+#         print('Saving merge database..')
+#         now = dt.datetime.now()
+#         db_name = now.strftime("Reid_Interrputed_%Y%m%d.db")
+#         db_filepath = os.path.join(args[3], db_name)
+#         reid_db = ImageDB(db_name=db_filepath)
+#         reid_db.delete_dbfile()
+#         reid_db.create_table()
+#         reid_db.merge_data(db_list)
 
-        sys.exit(0)
+#         sys.exit(0)
         
-    signal.signal(signal.SIGINT, signal_handler)
+#     signal.signal(signal.SIGINT, signal_handler)
 
-    while True:
-        db_list = [] 
-        # args[0] is the length of camera list
-        # args[1] is the shared queue recording the database paths 
-        while len(db_list) < args[1]:
-            print("db_path_process: ", args[4].get())
-            db_list.append(args[4].get())
-            #time.sleep(1)
-        # finish gathering the db_paths, run merge.
-        print('Saving merge database..')
-        now = dt.datetime.now()
-        db_name = now.strftime("Reid_%Y%m%d.db")
-        db_filepath = os.path.join(args[3], db_name)
-        reid_db = ImageDB(db_name=db_filepath)
-        reid_db.delete_dbfile()
-        reid_db.create_table()
-        reid_db.merge_data(db_list)
+#     while True:
+#         db_list = [] 
+#         # args[0] is the length of camera list
+#         # args[1] is the shared queue recording the database paths 
+#         while len(db_list) < args[1]:
+#             print("db_path_process: ", args[4].get())
+#             db_list.append(args[4].get())
+#             #time.sleep(1)
+#         # finish gathering the db_paths, run merge.
+#         print('Saving merge database..')
+#         now = dt.datetime.now()
+#         db_name = now.strftime("Reid_%Y%m%d.db")
+#         db_filepath = os.path.join(args[3], db_name)
+#         reid_db = ImageDB(db_name=db_filepath)
+#         reid_db.delete_dbfile()
+#         reid_db.create_table()
+#         reid_db.merge_data(db_list)
     
 
 
@@ -96,7 +97,7 @@ class MultiPs():
         self.cam = []
 
         # shared resource
-        self.db_queue = mp.Queue()
+        #self.db_queue = mp.Queue()
         self.manager = mp.Manager()
         self.unique_id = self.manager.list()
 
@@ -106,10 +107,10 @@ class MultiPs():
         logger.setLevel(logging.DEBUG)
 
     def new_job(self, name, target, *args):
-        print("args: ", args)
-        q_args = (*args, self.db_queue)
-        print("q_args: ", q_args)
-        j = mp.Process(name=name, target=target, args=q_args)
+        #print("args: ", args)
+        #q_args = (*args, self.db_queue)
+        #print("q_args: ", q_args)
+        j = mp.Process(name=name, target=target, args=args)
         j.daemon = True
         self.job.append(j)
 
@@ -121,11 +122,9 @@ class MultiPs():
     def signal_handler(self, sig, frame):
         print('Main Program: You pressed Ctrl+C!')
         # save db if the process is interrupted halfway.
-        for i in range(FLAGS.parallel_ps):
-            if self.cam:
-                self.db_queue.put(FLAGS.cam_db_path + "/Cam_" + str(self.cam[i]) + ".db")
-        # wait for dataase merging
-        time.sleep(10)
+        # for i in range(FLAGS.parallel_ps):
+        #     if self.cam:
+        #         self.db_queue.put(FLAGS.reid_db_path + "/Cam_" + str(self.cam[i]) + ".db")
         for j in self.job:
             j.join()
 
@@ -142,7 +141,7 @@ def cam_stream(mps):
 
 def sequential_run(batch, cam, db_path, mps):
     mps.job.clear()
-    mps.new_job('database_ps', db_process, cam, FLAGS.parallel_ps, FLAGS.cam_db_path, FLAGS.merge_db_path)
+    #mps.new_job('database_ps', db_process, cam, FLAGS.parallel_ps, FLAGS.reid_db_path, FLAGS.merge_db_path)
     mps.cam = cam
     print("batch:", batch)
     gpu_num = 0
@@ -156,7 +155,7 @@ def sequential_run(batch, cam, db_path, mps):
 
 def online_run(rtsp, cam, gpu, db_path, mps):
     mps.job.clear()
-    mps.new_job('database_ps', db_process, cam, FLAGS.parallel_ps, FLAGS.cam_db_path, FLAGS.merge_db_path)
+    #mps.new_job('database_ps', db_process, cam, FLAGS.parallel_ps, FLAGS.reid_db_path, FLAGS.merge_db_path)
     mps.cam = cam
     for i in range(FLAGS.parallel_ps):
         # cam[i]:int , rtsp[i]:str
@@ -213,13 +212,15 @@ def main(_argv):
     # mps.log_msg()
     print("Parent Process PID: " + str(os.getpid()))
     print("Initialize database..")
-
+    
     # initialize backup database
-    db_path = "./database/Image_" + str(dt.datetime.now().strftime("%Y%m%dT%H%M%S")) + ".db"
-    print("db_path main: ", db_path)
-    img_db = ImageDB(db_path)
-    img_db.delete_dbfile()
-    img_db.create_table()
+    db_path = None
+    if FLAGS.db:
+        db_path = "./database/Image_" + str(dt.datetime.now().strftime("%Y%m%dT%H%M%S")) + ".db"
+        print("db_path main: ", db_path)
+        img_db = ImageDB(db_path)
+        img_db.delete_dbfile()
+        img_db.create_table()
 
     # online mode
     if FLAGS.online:
